Removed test dimension files, made chunk provider random into non-static field.

+++ src/main/java/com/bespectacled/modernbeta/world/decorator/noise/AlphaNoiseDecorator.java
@@ -22,8 +22,7 @@ public class AlphaNoiseDecorator implements OldNoiseDecorator {
         
         double scale = 0.5D;
 
-        int noiseCount = (int) ((this.noiseSampler.sample((double) startX * scale, (double) startZ * scale) / 8D
-                + random.nextDouble() * 4D + 4D) / 3D);
+        int noiseCount = (int) ((this.noiseSampler.sample(startX * scale, startZ * scale) / 8D + random.nextDouble() * 4D + 4D) / 3D);
         
         if (noiseCount < 0)
             noiseCount = 0;
